Implement command processing
Implement the processing of the commands. A queue is used to store the
telecommands. On insert the command is added to the tail of the queue.
On execute the commands are removed from the head of the queue and
executed. A merge sort algorithm is used to sort the queue when a sort
command is received.

Signed-off-by: Todor Tomov <todor.too@gmail.com>